Initial work on multiple clock windows for FC Teleprinter
