+++ teleprinter/clockwindow.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>FightclubTeleprinter</class>
- <widget class="QMainWindow" name="FightclubTeleprinter">
+ <class>ClockWindow</class>
+ <widget class="QMainWindow" name="ClockWindow">
   <property name="geometry">
    <rect>
     <x>0</x>
@@ -17,7 +17,7 @@
    </size>
   </property>
   <property name="windowTitle">
-   <string>Fightclub Teleprinter</string>
+   <string>Fightclub Teleprinter Clock Window</string>
   </property>
   <property name="locale">
    <locale language="English" country="UnitedStates"/>
